\name{credit.data}
\docType{data}
\alias{credit.data}
\title{Statlog (German Credit Data) Data Set}
\description{
  This dataset classifies people described by a set of attributes as good or bad credit risks.
}
\usage{data(credit.data)}
\format{A vector containing 1000 observations.}
\source{UCI Repository, https://archive.ics.uci.edu/ml/datasets/Statlog+(German+Credit+Data).}
\keyword{datasets}