\name{ToDataFrame}
\alias{ToDataFrame}
\title{Transforms patterns into a data frame}
\description{
  Transforms a list/vector of patterns into a data frame
  for inspection and analysis.
}
\usage{
ToDataFrame(patterns, ndigits = 2)
}
\arguments{
\item{patterns}{List/vector of patterns.}
\item{ndigits}{Number of significant digits when printing floats (optional).}
}
\seealso{
\code{\link{DiscoverSubgroups}}.
}
\keyword{subgroup analysis}